ner_cli/requirements.txt
python-dotenv==1.0.1
Requests==2.31.0
tqdm==4.66.1
